update-thongKe

+++ QuanLyKhoApi/Controllers/HangHoaController.cs
@@ -3,18 +3,33 @@ using Microsoft.AspNetCore.Mvc;
 using QuanLyKhoApi.Dto;
 using QuanLyKhoApi.Helper;
 using QuanLyKhoApi.IServices;
+using System.Net;
 using System.Security.Claims;
 
 namespace QuanLyKhoApi.Controllers
 {
     [ApiController]
     [Route("api/[controller]")]
-    public class HangHoaController(IHangHoaService service) : ControllerBase
+    public class HangHoaController(IHangHoaService service, AuthorizationService auth) : ControllerBase
     {
+        private class ClaimHangHoa
+        {
+            public const string XemHangHoa = "XemHangHoa";
+            public const string ThemHangHoa = "ThemHangHoa";
+            public const string SuaHangHoa = "SuaHangHoa";
+            public const string XoaHangHoa = "XoaHangHoa";
+            public static readonly string[] HangHoaclaim = { XemHangHoa, ThemHangHoa, SuaHangHoa, XoaHangHoa };
+        }
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] HangHoaDto dto)
         {
+            var iduser = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isHasClaim = await auth.RoleHasClaimAsync(iduser, ClaimHangHoa.ThemHangHoa);
+            if (isHasClaim.Success == false)
+            {
+                return MyStatusCodeBase.MyStatusCode(this, isHasClaim);
+            }
             var result = await service.CreateHangHoaAsync(dto);
             return MyStatusCodeBase.MyStatusCode(this, result);
         }
@@ -26,6 +41,12 @@ namespace QuanLyKhoApi.Controllers
             [FromQuery] int pageSize = 12,
             [FromQuery] SortOBJ? sort = null)
         {
+            var iduser = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isHasClaim = await auth.RoleHasListClaimAsync(iduser, ClaimHangHoa.HangHoaclaim);
+            if (isHasClaim.Success == false)
+            {
+                return MyStatusCodeBase.MyStatusCode(this, isHasClaim);
+            }
             var result = await service.GetAllHangHoaAsync(query, page, pageSize, sort);
             return MyStatusCodeBase.MyStatusCode(this, result);
         }
@@ -33,6 +54,12 @@ namespace QuanLyKhoApi.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById(string id)
         {
+            var iduser = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isHasClaim = await auth.RoleHasListClaimAsync(iduser, ClaimHangHoa.HangHoaclaim);
+            if (isHasClaim.Success == false)
+            {
+                return MyStatusCodeBase.MyStatusCode(this, isHasClaim);
+            }
             var result = await service.GetHangHoaByIdAsync(id);
             return MyStatusCodeBase.MyStatusCode(this, result);
         }
@@ -40,6 +67,12 @@ namespace QuanLyKhoApi.Controllers
         [HttpPost("scan")]
         public async Task<IActionResult> GetByScanCode([FromForm] Getfile dto)
         {
+            var iduser = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isHasClaim = await auth.RoleHasListClaimAsync(iduser, ClaimHangHoa.HangHoaclaim);
+            if (isHasClaim.Success == false)
+            {
+                return MyStatusCodeBase.MyStatusCode(this, isHasClaim);
+            }
             var result = await service.FindByBarCode(dto.file);
             return MyStatusCodeBase.MyStatusCode(this, result);
         }
@@ -47,6 +80,12 @@ namespace QuanLyKhoApi.Controllers
         [HttpGet("{id}/gen-barcode")]
         public async Task<IActionResult> GenBarCode(string id)
         {
+            var iduser = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isHasClaim = await auth.RoleHasListClaimAsync(iduser, ClaimHangHoa.HangHoaclaim);
+            if (isHasClaim.Success == false)
+            {
+                return MyStatusCodeBase.MyStatusCode(this, isHasClaim);
+            }
             var result = await service.GenBarCode(id);
             return MyStatusCodeBase.MyStatusCode(this, result);
         }
@@ -54,6 +93,12 @@ namespace QuanLyKhoApi.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> Update(Guid id, [FromBody] HangHoaDto dto)
         {
+            var iduser = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isHasClaim = await auth.RoleHasClaimAsync(iduser, ClaimHangHoa.SuaHangHoa);
+            if (isHasClaim.Success == false)
+            {
+                return MyStatusCodeBase.MyStatusCode(this, isHasClaim);
+            }
             var result = await service.UpdateHangHoaAsync(id, dto);
             return MyStatusCodeBase.MyStatusCode(this, result);
         }
@@ -61,6 +106,12 @@ namespace QuanLyKhoApi.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(Guid id)
         {
+            var iduser = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isHasClaim = await auth.RoleHasClaimAsync(iduser, ClaimHangHoa.XoaHangHoa);
+            if (isHasClaim.Success == false)
+            {
+                return MyStatusCodeBase.MyStatusCode(this, isHasClaim);
+            }
             var result = await service.DeleteHangHoaAsync(id);
             return MyStatusCodeBase.MyStatusCode(this, result);
         }
@@ -68,6 +119,12 @@ namespace QuanLyKhoApi.Controllers
         [HttpGet("{cauHinhId}/configs")]
         public async Task<IActionResult> GetConfigs(Guid hangHoaId)
         {
+            var iduser = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isHasClaim = await auth.RoleHasListClaimAsync(iduser, ClaimHangHoa.HangHoaclaim);
+            if (isHasClaim.Success == false)
+            {
+                return MyStatusCodeBase.MyStatusCode(this, isHasClaim);
+            }
             var result = await service.GetCauHinhById(hangHoaId);
             return MyStatusCodeBase.MyStatusCode(this, result);
         }
@@ -75,6 +132,12 @@ namespace QuanLyKhoApi.Controllers
         [HttpPost("configs/{id}")]
         public async Task<IActionResult> CreateMultipleConfigs([FromRoute] Guid id, [FromBody] List<CreateCauHinhDto> dto)
         {
+            var iduser = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isHasClaim = await auth.RoleHasClaimAsync(iduser, ClaimHangHoa.ThemHangHoa);
+            if (isHasClaim.Success == false)
+            {
+                return MyStatusCodeBase.MyStatusCode(this, isHasClaim);
+            }
             var result = await service.CreateMultipleConfigsAsync(id, dto);
             return MyStatusCodeBase.MyStatusCode(this, result);
         }
@@ -82,6 +145,12 @@ namespace QuanLyKhoApi.Controllers
         [HttpPost("configs/{id}/images")]
         public async Task<IActionResult> AddHinhAnhCauHing([FromRoute] Guid id, [FromForm] IFormFile[] files)
         {
+            var iduser = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isHasClaim = await auth.RoleHasClaimAsync(iduser, ClaimHangHoa.SuaHangHoa);
+            if (isHasClaim.Success == false)
+            {
+                return MyStatusCodeBase.MyStatusCode(this, isHasClaim);
+            }
             var result = await service.AddHinhAnhCauHing(id, files);
             return MyStatusCodeBase.MyStatusCode(this, result);
         }
@@ -89,6 +158,12 @@ namespace QuanLyKhoApi.Controllers
         [HttpDelete("configs/images")]
         public async Task<IActionResult> DeleteHinhAnhCauHinh([FromBody] Guid[] id)
         {
+            var iduser = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var isHasClaim = await auth.RoleHasClaimAsync(iduser, ClaimHangHoa.SuaHangHoa);
+            if (isHasClaim.Success == false)
+            {
+                return MyStatusCodeBase.MyStatusCode(this, isHasClaim);
+            }
             var result = await service.DeleteHinhAnhCauHinhAsync(id);
             return MyStatusCodeBase.MyStatusCode(this, result);
         }
